Refactored Bedrock.Directory
